Move the test to MessageIdSerializationTest.java
